# ggml-cpu: faster AVX2 variant for IQ1_M