This is an additional optimization for AVX2 after #12154

Benchmarks on AMD Zen 4:
```
Before
  MUL_MAT(type_a=iq1_m,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  1932 runs -   532.51 us/run - 117.44 MFLOP/run - 220.54 GFLOPS
  MUL_MAT(type_a=iq1_m,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                   980 runs -  1057.22 us/run - 234.88 MFLOP/run - 222.17 GFLOPS

After
  MUL_MAT(type_a=iq1_m,type_b=f32,m=4096,n=1,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  2553 runs -   394.78 us/run - 117.44 MFLOP/run - 297.48 GFLOPS
  MUL_MAT(type_a=iq1_m,type_b=f32,m=4096,n=2,k=14336,bs=[1,1],nr=[1,1],per=[0,1,2,3]):                  1225 runs -   816.50 us/run - 234.88 MFLOP/run - 287.67 GFLOPS
```

| model                          |       size |     params | ngl |          test |  master          t/s | PR               t/s |                   
| ------------------------------ | ---------: | ---------: | --: | ------------: | -------------------: | -------------------: |
| qwen2 7B IQ1_M - 1.75 bpw      |   1.90 GiB |     7.62 B |   0 |         pp512 |        204.84 ± 0.31 |        204.07 ± 0.58 |
| qwen2 7B IQ1_M - 1.75 bpw      |   1.90 GiB |     7.62 B |   0 |         tg128 |         14.48 ± 0.02 |         17.71 ± 0.18 |